objhits: CheckObjectHitVolumes 95.55->95.88 ((uint) launders on the attA/attB null tests -- cmplwi per #3/#14; unit 96.96->97.0)
Residual: whole-fn #108 saved-reg rotation (params homed r29/r30/r21/r31
in target vs our contiguous r27-r29; state locals permuted) -- the giant
difflib regions are alignment artifacts of the reg numbers, instruction
order is parallel. ObjHits_Update parked at 92.75: the while+inline-
increment continue form matched the skip-path shape but net-regressed
(-0.22, reverted); opt_strength_reduction off inert on the fused
base+scaled walker; (i << 2) byte-offset spelling did not unfold the
clear loop against the symbol base (#112(c)).
Zero regressions, dol intact

+++ src/main/objhits.c
@@ -1687,12 +1687,12 @@ void ObjHits_CheckObjectHitVolumes(int objA,int objB,int attA,int attB,f32 dt)
 
   stateA = (ObjHitsPriorityState *)((GameObject *)objA)->anim.hitReactState;
   stateB = (ObjHitsPriorityState *)((GameObject *)objB)->anim.hitReactState;
-  if (attA != 0) {
+  if ((uint)attA != 0) {
     attStateA = *(ObjHitsPriorityState **)(attA + 0x54);
   } else {
     attStateA = NULL;
   }
-  if (attB != 0) {
+  if ((uint)attB != 0) {
     attStateB = *(ObjHitsPriorityState **)(attB + 0x54);
   } else {
     attStateB = NULL;
